requirements.txt
streamlit==1.39.0
databricks-sdk==0.34.0
requests==2.32.3